// src/lib/types.ts
export type Colors = {
  "light/primary": string;
  "light/gray": string;
  "light/background": string;
  "dark/primary": string;
  "dark/gray": string;
  "dark/background": string;
};

export type ShadcnCssVariables = {
  background?: string;
  foreground?: string;
  card?: string;
  cardForeground?: string;
  popover?: string;
  popoverForeground?: string;
  primary?: string;
  primaryForeground?: string;
  secondary?: string;
  secondaryForeground?: string;
  muted?: string;
  mutedForeground?: string;
  accent?: string;
  accentForeground?: string;
  destructive?: string;
  destructiveForeground?: string;
  border?: string;
  input?: string;
  ring?: string;
  chart1?: string;
  chart2?: string;
  chart3?: string;
  chart4?: string;
  chart5?: string;
  radius?: string;
};

export type ColorPalette = {
  label: string;
  value: string;
  example?: string;
};